+++ Sourcery/Generating/Template/Swift/SwiftTemplate.swift
@@ -9,11 +9,6 @@
 import Foundation
 import PathKit
 
-fileprivate enum SwiftTemplateParsingError: Error {
-    case unmatchedOpening(path: Path, line: Int)
-    case compilationError(path: Path, error: String)
-}
-
 fileprivate enum Delimiters {
     static let open = "<%"
     static let close = "%>"
@@ -25,6 +20,7 @@ fileprivate struct ProcessResult {
 }
 
 class SwiftTemplate: Template {
+
     let sourcePath: Path
 
     init(path: Path) throws {
@@ -51,7 +47,7 @@ class SwiftTemplate: Template {
 
         for component in components.suffix(from: 1) {
             guard let endIndex = component.range(of: Delimiters.close) else {
-                throw SwiftTemplateParsingError.unmatchedOpening(path: path, line: currentLineNumber())
+                throw "\(path):\(currentLineNumber()) Error while parsing template. Unmatched <%"
             }
 
             var code = component.substring(to: endIndex.lowerBound)
@@ -154,11 +150,15 @@ class SwiftTemplate: Template {
                 let command = "/usr/bin/swiftc " + arguments.map { "\"\($0)\"" }.joined(separator: " ")
                 print(command)
             #endif
-            throw SwiftTemplateParsingError.compilationError(path: binaryFile, error: compilationResult.error)
+            throw compilationResult.error
         }
 
         let result = try Process.runCommand(path: binaryFile.description,
                                             arguments: [serializedContextPath.description])
+
+        if !result.error.isEmpty {
+            throw "\(sourcePath): \(result.error)"
+        }
 
         return result.output
     }
@@ -171,17 +171,6 @@ fileprivate extension SwiftTemplate {
 
     static var swiftTemplatesRuntime: Path {
         return resourcesPath + Path("SwiftTemplateRuntime")
-    }
-}
-
-extension SwiftTemplateParsingError: CustomStringConvertible {
-    var description: String {
-        switch self {
-        case let .unmatchedOpening(path, line):
-            return "\(path):\(line) Error while parsing template. Unmatched <%"
-        case let .compilationError(_, error):
-            return error
-        }
     }
 }