+++ SourceryTests/Parsing/FileParser + MethodsSpec.swift
@@ -15,7 +15,7 @@ class FileParserMethodsSpec: QuickSpec {
             describe("parseMethod") {
                 func parse(_ code: String) -> [Type] {
                     guard let parserResult = try? FileParser(contents: code).parse() else { fail(); return [] }
-                    return Composer(verbose: false).uniqueTypes(parserResult)
+                    return Composer().uniqueTypes(parserResult)
                 }
 
                 it("extracts methods properly") {