+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### New Features
 
 - Paths in config file are now relative to config file path by default, absolute paths should start with `/`
+- Improved logging and error reporting, added `--quiet` CLI option, added runtime errors for using invalid types in `implementing` and `inheriting`
 
 
 ## 0.6.0